# frontend/openapi.json
{
  "components": {
    "schemas": {
      "ActionId": {
        "description": "Policy §8. ACT-06 (prohibited actions) is intentionally absent: it is never authorized.",
        "enum": [
          "ACT-01",
          "ACT-02",
          "ACT-03",
          "ACT-04",
          "ACT-05"
        ],
        "title": "ActionId",
        "type": "string"
      },
      "ActionTaken": {
        "additionalProperties": false,
        "description": "An attempted action with its verification result. Failed actions are included.",
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ActionId"
          },
          "at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "At"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "result": {
            "$ref": "#/components/schemas/ToolStatus"
          },
          "verified": {
            "title": "Verified",
            "type": "boolean"
          }
        },
        "required": [
          "action",
          "result",
          "verified"
        ],
        "title": "ActionTaken",
        "type": "object"
      },
...
      "AuthStatus": {
        "enum": [
          "authenticated",
          "unauthenticated",
          "expired"
        ],
        "title": "AuthStatus",
        "type": "string"
      },
...
      "CollectedSlot": {
        "additionalProperties": false,
        "description": "A slot the customer already gave: a claim, never verified (DATA-03).",
        "properties": {
          "name": {
            "$ref": "#/components/schemas/SlotName"
          },
          "turn_index": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turn Index"
          },
          "value": {
            "minLength": 1,
            "title": "Value",
            "type": "string"
          },
          "verified": {
            "const": false,
            "default": false,
            "title": "Verified",
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "value"
        ],
        "title": "CollectedSlot",
        "type": "object"
      },
      "DraftCase": {
        "additionalProperties": false,
        "properties": {
          "amount_usd": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount Usd"
          },
          "provisional_credit_flag": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProvisionalCreditFlag"
              },
              {
                "type": "null"
              }
            ]
          },
          "tier": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Tier"
              },
              {
                "type": "null"
              }
            ]
          },
          "transaction_ref": {
            "minLength": 1,
            "title": "Transaction Ref",
            "type": "string"
          }
        },
        "required": [
          "transaction_ref"
        ],
        "title": "DraftCase",
        "type": "object"
      },
      "EscalationReason": {
        "additionalProperties": false,
        "description": "A triggered rule in plain words, with the evidence that made it fire.",
        "properties": {
          "description": {
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "minItems": 1,
            "title": "Evidence",
            "type": "array"
          },
          "rule_id": {
            "pattern": "^(GATE|ESC)-\\d{2}$",
            "title": "Rule Id",
            "type": "string"
          }
        },
        "required": [
          "rule_id",
          "description",
          "evidence"
        ],
        "title": "EscalationReason",
        "type": "object"
      },
      "Evidence": {
        "additionalProperties": false,
        "description": "One input that made a rule fire. Taken from the request, never generated.",
        "properties": {
          "claims": {
            "items": {
              "minLength": 1,
              "type": "string"
            },
            "title": "Claims",
            "type": "array"
          },
          "kind": {
            "$ref": "#/components/schemas/EvidenceKind"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "origin": {
            "minLength": 1,
            "title": "Origin",
            "type": "string"
          },
          "record_id": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Record Id"
          },
          "source": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "value": {
            "minLength": 1,
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "name",
          "value",
          "origin"
        ],
        "title": "Evidence",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HandoffAuth": {
        "additionalProperties": false,
        "properties": {
          "method": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Method"
          },
          "session_age_min": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Age Min"
          },
          "status": {
            "$ref": "#/components/schemas/AuthStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "HandoffAuth",
        "type": "object"
      },
      "HandoffModelSignals": {
        "additionalProperties": false,
        "description": "Informative model output shown to the agent (never a decision).",
        "properties": {
          "calibrated": {
            "default": false,
            "title": "Calibrated",
            "type": "boolean"
          },
          "escalation_risk": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Escalation Risk"
          },
          "model_info": {
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "minLength": 1
            },
            "title": "Model Info",
            "type": "object"
          },
          "model_version": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "reason_code_other": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Code Other"
          },
          "reason_code_probs": {
            "additionalProperties": {
              "maximum": 1.0,
              "minimum": 0.0,
              "type": "number"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/ReasonCode"
            },
            "title": "Reason Code Probs",
            "type": "object"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelSource"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "HandoffModelSignals",
        "type": "object"
      },
      "HandoffPacket": {
        "additionalProperties": false,
        "description": "Policy §13. Never contains the raw transcript, only ``transcript_ref``.",
        "properties": {
          "actions_taken": {
            "items": {
              "$ref": "#/components/schemas/ActionTaken"
            },
            "title": "Actions Taken",
            "type": "array"
          },
          "auth": {
            "$ref": "#/components/schemas/HandoffAuth"
          },
          "business_date": {
            "format": "date",
            "title": "Business Date",
            "type": "string"
          },
          "collected_slots": {
            "items": {
              "$ref": "#/components/schemas/CollectedSlot"
            },
            "title": "Collected Slots",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "customer_claims": {
            "items": {
              "minLength": 1,
              "type": "string"
            },
            "title": "Customer Claims",
            "type": "array"
          },
          "customer_ref": {
            "minLength": 1,
            "title": "Customer Ref",
            "type": "string"
          },
          "draft_case": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DraftCase"
              },
              {
                "type": "null"
              }
            ]
          },
          "escalation_reasons": {
            "items": {
              "$ref": "#/components/schemas/EscalationReason"
            },
            "title": "Escalation Reasons",
            "type": "array"
          },
          "handoff_id": {
            "pattern": "^HO-\\d{8}-\\d{6,}$",
            "title": "Handoff Id",
            "type": "string"
          },
          "language": {
            "$ref": "#/components/schemas/Language"
          },
          "model_signals": {
            "$ref": "#/components/schemas/HandoffModelSignals"
          },
          "open_questions": {
            "items": {
              "minLength": 1,
              "type": "string"
            },
            "title": "Open Questions",
            "type": "array"
          },
          "policy_version": {
            "minLength": 1,
            "title": "Policy Version",
            "type": "string"
          },
          "post_handoff_messages": {
            "items": {
              "$ref": "#/components/schemas/PostHandoffMessage"
            },
            "title": "Post Handoff Messages",
            "type": "array"
          },
          "priority": {
            "$ref": "#/components/schemas/Priority"
          },
          "queue": {
            "$ref": "#/components/schemas/Queue"
          },
          "reason_code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReasonCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_summary": {
            "minLength": 1,
            "title": "Request Summary",
            "type": "string"
          },
          "transcript_ref": {
            "minLength": 1,
            "title": "Transcript Ref",
            "type": "string"
          },
          "triggered_rules": {
            "items": {
              "pattern": "^(GATE|ESC)-\\d{2}$",
              "type": "string"
            },
            "minItems": 1,
            "title": "Triggered Rules",
            "type": "array"
          },
          "verified_facts": {
            "items": {
              "$ref": "#/components/schemas/VerifiedFact"
            },
            "title": "Verified Facts",
            "type": "array"
          }
        },
        "required": [
          "handoff_id",
          "created_at",
          "business_date",
          "language",
          "queue",
          "priority",
          "customer_ref",
          "auth",
          "request_summary",
          "triggered_rules",
          "transcript_ref",
          "policy_version"
        ],
        "title": "HandoffPacket",
        "type": "object"
      },
      "HandoffPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/HandoffSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "title": "HandoffPage",
        "type": "object"
      },
      "HandoffSummary": {
        "additionalProperties": false,
        "description": "One line of the queue: enough to choose a case without opening it.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "handoff_id": {
            "title": "Handoff Id",
            "type": "string"
          },
          "language": {
            "$ref": "#/components/schemas/Language"
          },
          "priority": {
            "$ref": "#/components/schemas/Priority"
          },
          "queue": {
            "$ref": "#/components/schemas/Queue"
          },
          "request_summary": {
            "title": "Request Summary",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "triggered_rules": {
            "items": {
              "type": "string"
            },
            "title": "Triggered Rules",
            "type": "array"
          }
        },
        "required": [
          "handoff_id",
          "created_at",
          "queue",
          "priority",
          "status",
          "language",
          "request_summary",
          "triggered_rules"
        ],
        "title": "HandoffSummary",
        "type": "object"
      },
...
      "Language": {
        "description": "Supported conversation languages (GATE-01, COM-01).",
        "enum": [
          "es",
          "pt"
        ],
        "title": "Language",
        "type": "string"
      },
...
      "ModelSource": {
        "description": "Where decision-layer signals came from (architecture §3, §8).",
        "enum": [
          "kev",
          "llm_fallback",
          "unavailable"
        ],
        "title": "ModelSource",
        "type": "string"
      },
      "Outcome": {
        "description": "Policy §9. Precedence: REFUSE > ESCALATE > INFORM > CLARIFY > RESOLVE.",
        "enum": [
          "RESOLVE",
          "CLARIFY",
          "INFORM",
          "ESCALATE",
          "REFUSE"
        ],
        "title": "Outcome",
        "type": "string"
      },
...
      "PostHandoffMessage": {
        "additionalProperties": false,
        "description": "A message the customer wrote after the handoff, masked, read by no model.",
        "properties": {
          "received_at": {
            "format": "date-time",
            "title": "Received At",
            "type": "string"
          },
          "text": {
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "received_at",
          "text"
        ],
        "title": "PostHandoffMessage",
        "type": "object"
      },
      "Priority": {
        "enum": [
          "normal",
          "high"
        ],
        "title": "Priority",
        "type": "string"
      },
      "ProvisionalCreditFlag": {
        "description": "Policy §6. For T3 the flag is decided by a human, so it is absent (None).",
        "enum": [
          "eligible",
          "requires_review"
        ],
        "title": "ProvisionalCreditFlag",
        "type": "string"
      },
      "Queue": {
        "description": "Escalation routes named in policy §7.",
        "enum": [
          "disputes",
          "fraud",
          "security_review"
        ],
        "title": "Queue",
        "type": "string"
      },
      "ReasonCode": {
        "description": "Policy §3.",
        "enum": [
          "RC_UNRECOGNIZED",
          "RC_DUPLICATE",
          "RC_INCORRECT_AMOUNT",
          "RC_NOT_RECEIVED",
          "RC_FEE"
        ],
        "title": "ReasonCode",
        "type": "string"
      },
...
      "SlotName": {
        "description": "Policy §10, in the order the system asks for them.",
        "enum": [
          "transaction_ref",
          "reason_code",
          "card_in_possession",
          "shared_credentials",
          "duplicate_ref",
          "expected_amount",
          "expected_delivery_date",
          "merchant_contacted",
          "fee_ref",
          "confirmation"
        ],
        "title": "SlotName",
        "type": "string"
      },
      "Tier": {
        "description": "Policy §6.",
        "enum": [
          "T1",
          "T2",
          "T3"
        ],
        "title": "Tier",
        "type": "string"
      },
...
      "ToolStatus": {
        "enum": [
          "success",
          "failed",
          "access_denied"
        ],
        "title": "ToolStatus",
        "type": "string"
      },
      "TracePage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TraceSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "title": "TracePage",
        "type": "object"
      },
...
      "TraceSummary": {
        "additionalProperties": false,
        "description": "One line of the console's trace list: enough to choose a turn without loading it.",
        "properties": {
          "conversation_id": {
            "title": "Conversation Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "estimated_cost_usd": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Cost Usd"
          },
          "handoff_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Handoff Id"
          },
          "language": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Language"
              },
              {
                "type": "null"
              }
            ]
          },
          "outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Outcome"
              },
              {
                "type": "null"
              }
            ]
          },
          "reply_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reply Kind"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "total_latency_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Latency Ms"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          },
          "turn_index": {
            "title": "Turn Index",
            "type": "integer"
          }
        },
        "required": [
          "trace_id",
          "conversation_id",
          "session_id",
          "turn_index",
          "created_at",
          "language",
          "outcome",
          "reply_kind",
          "handoff_id",
          "total_latency_ms",
          "estimated_cost_usd"
        ],
        "title": "TraceSummary",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VerifiedFact": {
        "additionalProperties": false,
        "description": "DATA-04: every verified fact names its source table and record ID.",
        "properties": {
          "fact": {
            "minLength": 1,
            "title": "Fact",
            "type": "string"
          },
          "record_id": {
            "minLength": 1,
            "title": "Record Id",
            "type": "string"
          },
          "source": {
            "minLength": 1,
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "fact",
          "source",
          "record_id"
        ],
        "title": "VerifiedFact",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Dispute Intake API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/agent/handoffs": {
      "get": {
        "description": "Escalated cases, high priority first, then the oldest. ``search`` matches part of the\nhandoff ID the customer was given (HO-...), the conversation ID, the customer reference or the\nrequest summary.",
        "operationId": "list_handoffs_api_agent_handoffs_get",
        "parameters": [
          {
            "in": "query",
            "name": "queue",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Queue"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Queue"
            }
          },
          {
            "in": "query",
            "name": "priority",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Priority"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority"
            }
          },
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandoffPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Handoffs",
        "tags": [
          "agent"
        ]
      }
    },
    "/api/agent/handoffs/{handoff_id}": {
      "get": {
        "operationId": "get_handoff_api_agent_handoffs__handoff_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "handoff_id",
            "required": true,
            "schema": {
              "title": "Handoff Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandoffPacket"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Handoff",
        "tags": [
          "agent"
        ]
      }
    },
...
    "/api/agent/session": {
      "get": {
        "description": "204 if the token proves the agent role, 403 otherwise: the console's sign-in check. Recorded\nlike any other read.",
        "operationId": "check_session_api_agent_session_get",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Check Session",
        "tags": [
          "agent"
        ]
      }
    },
    "/api/agent/traces": {
      "get": {
        "description": "Turn traces, newest first, as summaries. ``search`` matches part of a trace,\nconversation, session or handoff ID; the full trace is read from /api/audit/{trace_id}.",
        "operationId": "search_traces_api_agent_traces_get",
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "outcome",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Outcome"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Outcome"
            }
          },
          {
            "in": "query",
            "name": "language",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Language"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Language"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TracePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Search Traces",
        "tags": [
          "agent"
        ]
      }
    },
...
  }
}